♻️ Mv getTransactions to Visualization component
