+++ src/types/index.ts
@@ -1,0 +1,7 @@
+export interface Transaction {
+    time: Date;
+    amount: string;
+    type: string;
+    subtype: string;
+    link: string;
+}
